logger/requirements.txt
pynamodb
pyyaml
